<fix>[alerts]:Unified alerts in a single file
